+++ cookbooks/linux/metadata.rb
@@ -1,7 +1,0 @@
-name             'linux'
-maintainer       'Replica Labs'
-maintainer_email 'jonschwadron@gmail.com'
-license          'All rights reserved'
-description      'Installs/Configures linux'
-long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.0'