+++ cookbooks/replicalabs/CHANGELOG.md
@@ -5,9 +5,17 @@ This file is used to list changes made in each version of the replicalabs cookbo
 
 0.1.0
 -----
-- [your_name] - Initial release of replicalabs
+- [Jonathan Schwadron] - Initial release of replicalabs
 
 - - -
+
+0.1.1
+-----
+- [Jonathan Schwadron] - Merged mutliple recipes from other cookbooks into one cookbook.
+
+- - -
+
+
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.
 
 The [Github Flavored Markdown page](http://github.github.com/github-flavored-markdown/) describes the differences between markdown on github and standard markdown.